feat: Integrate vibe-json-editor into local-storage-editor for enhanced JSON editing capabilities
- Added visual JSON editing functionality to local-storage-editor.
- Removed local-storage-manager.js as its functionality is now integrated into local-storage-editor.
- Updated vibe-json-editor with improved styling and functionality for better user experience.
- Created a new test page to showcase various web components including the updated local-storage-editor.
