+++ experiments/wc/test-components.html
@@ -1,0 +1,139 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Web Components Test Page</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            margin: 0;
+            padding: 20px;
+            background-color: #f5f5f5;
+        }
+        .container {
+            max-width: 1200px;
+            margin: 0 auto;
+        }
+        h1 {
+            text-align: center;
+            color: #333;
+        }
+        .component-section {
+            margin-bottom: 40px;
+            padding: 20px;
+            background: white;
+            border-radius: 8px;
+            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+        }
+        .component-section h2 {
+            margin-top: 0;
+            color: #555;
+            border-bottom: 2px solid #eee;
+            padding-bottom: 10px;
+        }
+        .grid {
+            display: grid;
+            grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+            gap: 20px;
+        }
+        .component-item {
+            padding: 15px;
+            border: 1px solid #ddd;
+            border-radius: 4px;
+            background: #fafafa;
+        }
+        .component-item h3 {
+            margin-top: 0;
+            color: #777;
+            font-size: 1em;
+        }
+    </style>
+
+    <!-- Scripts for web components -->
+    <script type="module" src="color-box.js"></script>
+    <script type="module" src="counter-card.js"></script>
+    <script type="module" src="my-card.js"></script>
+    <script type="module" src="local-storage-editor.js"></script>
+    <script src="notes-panel.js"></script>
+    <script src="siri-prompt-interface.js"></script>
+    <script src="smart-chart.js"></script>
+    <script src="task-list.js"></script>
+    <script src="vibe-timer-alarm.js"></script>
+</head>
+<body>
+    <div class="container">
+        <h1>Web Components Test Suite</h1>
+
+        <div class="component-section">
+            <h2>Basic Components</h2>
+            <div class="grid">
+                <div class="component-item">
+                    <h3>Color Box</h3>
+                    <color-box color="red"></color-box>
+                    <color-box color="#00ff00"></color-box>
+                    <color-box color="blue"></color-box>
+                </div>
+                <div class="component-item">
+                    <h3>Counter Card</h3>
+                    <counter-card value="5"></counter-card>
+                    <counter-card value="42"></counter-card>
+                </div>
+                <div class="component-item">
+                    <h3>My Card</h3>
+                    <my-card></my-card>
+                </div>
+            </div>
+        </div>
+
+        <div class="component-section">
+            <h2>Timer and Alarm</h2>
+            <div class="grid">
+                <div class="component-item">
+                    <h3>Vibe Timer Alarm</h3>
+                    <vibe-timer-alarm countdown-duration-seconds="10" accent-color-hex="#ff6b6b"></vibe-timer-alarm>
+                </div>
+            </div>
+        </div>
+
+        <div class="component-section">
+            <h2>Charts and Lists</h2>
+            <div class="grid">
+                <div class="component-item">
+                    <h3>Smart Chart</h3>
+                    <smart-chart chart-data-values="[10,20,30,40,50]" chart-display-label="Sample Data"></smart-chart>
+                </div>
+                <div class="component-item">
+                    <h3>Task List</h3>
+                    <task-list></task-list>
+                </div>
+            </div>
+        </div>
+
+        <div class="component-section">
+            <h2>Storage Editors</h2>
+            <div class="grid">
+                <div class="component-item">
+                    <h3>Local Storage Editor</h3>
+                    <local-storage-editor accent-color-hex="#4ecdc4" editor-height-pixels="300"></local-storage-editor>
+                </div>
+            </div>
+        </div>
+
+
+        <div class="component-section">
+            <h2>Notes and Interface</h2>
+            <div class="grid">
+                <div class="component-item">
+                    <h3>Notes Panel</h3>
+                    <notes-panel note="First note" note="Second note"></notes-panel>
+                </div>
+                <div class="component-item">
+                    <h3>Siri Prompt Interface</h3>
+                    <siri-prompt-interface input-placeholder-text="Type your prompt here..." glow-aura-color-scheme="blue"></siri-prompt-interface>
+                </div>
+            </div>
+        </div>
+    </div>
+</body>
+</html>
