+++ experiments/wc/vibe-json-editor.js
@@ -1,7 +1,9 @@
+// Code for json-editor.js
 
                             /**
  * <vibe-json-editor>
  * A high-quality, standalone graphical JSON editor web component.
+ * Features: Collapsible nodes, type-sensitive inputs, live editing, and reactive attributes.
  */
 customElements.define('vibe-json-editor', class extends HTMLElement {
   static get observedAttributes() {
@@ -24,12 +26,15 @@ customElements.define('vibe-json-editor', class extends HTMLElement {
 
   attributeChangedCallback(name, oldVal, newVal) {
     if (oldVal === newVal) return;
+    
     switch (name) {
       case 'json-string-content':
         try {
           this._data = JSON.parse(newVal || '{}');
           this._refreshTree();
-        } catch (e) {}
+        } catch (e) {
+          console.error('vibe-json-editor: Invalid JSON string provided.', e);
+        }
         break;
       case 'accent-ui-color':
         this._accentColor = newVal;
@@ -48,7 +53,9 @@ customElements.define('vibe-json-editor', class extends HTMLElement {
 
   _updateFromAttribute() {
     const attrJson = this.getAttribute('json-string-content');
-    if (attrJson) { try { this._data = JSON.parse(attrJson); } catch(e) {} }
+    if (attrJson) {
+      try { this._data = JSON.parse(attrJson); } catch(e) {}
+    }
     this._refreshTree();
   }
 
@@ -62,33 +69,95 @@ customElements.define('vibe-json-editor', class extends HTMLElement {
 
   _updateStyles() {
     const styleTag = this.shadowRoot.querySelector('style');
-    if (styleTag) styleTag.textContent = this._generateCSS();
+    if (styleTag) {
+      styleTag.textContent = this._generateCSS();
+    }
   }
 
   _generateCSS() {
     return `
       :host {
-        display: block; font-family: 'Segoe UI', system-ui, sans-serif;
-        --accent: ${this._accentColor}; --bg: #ffffff; --text: #1e293b; --border: #e2e8f0;
-        --key: #7c3aed; --string: #059669; --number: #d97706; --bool: #2563eb;
-        background: var(--bg); font-size: ${this._fontSize}; color: var(--text);
+        display: block;
+        font-family: 'Segoe UI', system-ui, -apple-system, sans-serif;
+        --accent: ${this._accentColor};
+        --bg: #ffffff;
+        --text: #1e293b;
+        --border: #e2e8f0;
+        --key: #7c3aed;
+        --string: #059669;
+        --number: #d97706;
+        --bool: #2563eb;
+        background: var(--bg);
+        border: 1px solid var(--border);
+        border-radius: 8px;
+        padding: 1rem;
+        font-size: ${this._fontSize};
+        color: var(--text);
+        box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1);
       }
-      .tree-node { margin-left: 1.2rem; position: relative; border-left: 1px solid var(--border); }
+      .tree-node { margin-left: 1.5rem; position: relative; }
+      .tree-node::before {
+        content: "";
+        position: absolute;
+        left: -10px;
+        top: 0;
+        bottom: 0;
+        width: 1px;
+        background: var(--border);
+      }
       .row { display: flex; align-items: center; padding: 2px 0; gap: 8px; }
-      .collapser { cursor: pointer; user-select: none; width: 16px; text-align: center; }
+      .collapser { 
+        cursor: pointer; 
+        user-select: none; 
+        width: 16px; 
+        display: inline-flex; 
+        justify-content: center;
+        transition: transform 0.2s;
+      }
       .collapsed .children { display: none; }
       .collapsed .collapser { transform: rotate(-90deg); }
       .key { color: var(--key); font-weight: 600; }
-      input { border: 1px transparent; background: transparent; font-family: inherit; font-size: inherit; color: inherit; padding: 1px 4px; border-radius: 4px; }
-      input:focus { outline: none; background: #fff; border-color: var(--accent); }
-      .actions { opacity: 0; display: flex; gap: 2px; }
+      .val-string { color: var(--string); }
+      .val-number { color: var(--number); }
+      .val-boolean { color: var(--bool); }
+      input {
+        border: 1px transparent;
+        background: transparent;
+        font-family: inherit;
+        font-size: inherit;
+        color: inherit;
+        padding: 2px 4px;
+        border-radius: 4px;
+        transition: all 0.2s;
+      }
+      input:hover:not([disabled]) { background: #f1f5f9; border-color: var(--border); }
+      input:focus { outline: none; background: #fff; border-color: var(--accent); box-shadow: 0 0 0 2px rgba(99, 102, 241, 0.2); }
+      .actions { opacity: 0; transition: opacity 0.2s; display: flex; gap: 4px; }
       .row:hover .actions { opacity: 1; }
-      .btn { cursor: pointer; border: none; background: #f1f5f9; border-radius: 4px; width: 18px; height: 18px; font-size: 10px; }
+      .btn {
+        cursor: pointer;
+        border: none;
+        background: #f1f5f9;
+        border-radius: 4px;
+        width: 20px;
+        height: 20px;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        font-size: 12px;
+        color: #64748b;
+      }
+      .btn:hover { background: #e2e8f0; color: #0f172a; }
+      .btn-del:hover { background: #fee2e2; color: #ef4444; }
+      .type-label { font-size: 10px; text-transform: uppercase; opacity: 0.5; font-weight: bold; }
     `;
   }
 
   _renderSkeleton() {
-    this.shadowRoot.innerHTML = `<style>${this._generateCSS()}</style><div id="container"></div>`;
+    this.shadowRoot.innerHTML = `
+      <style>${this._generateCSS()}</style>
+      <div id="container"></div>
+    `;
   }
 
   _refreshTree() {
@@ -101,255 +170,160 @@ customElements.define('vibe-json-editor', class extends HTMLElement {
   _createNode(key, value, parentObj, isRoot = false) {
     const wrapper = document.createElement('div');
     wrapper.className = 'tree-node';
+    
     const isObject = value !== null && typeof value === 'object';
     const row = document.createElement('div');
     row.className = 'row';
 
+    // Collapse toggle for objects
     if (isObject) {
-      const coll = document.createElement('span');
-      coll.className = 'collapser';
-      coll.textContent = '▼';
-      coll.onclick = () => wrapper.classList.toggle('collapsed');
-      row.appendChild(coll);
+      const collapser = document.createElement('span');
+      collapser.className = 'collapser';
+      collapser.textContent = '▼';
+      collapser.onclick = () => wrapper.classList.toggle('collapsed');
+      row.appendChild(collapser);
     } else {
-      const sp = document.createElement('span');
-      sp.className = 'collapser';
-      row.appendChild(sp);
+      const spacer = document.createElement('span');
+      spacer.className = 'collapser';
+      row.appendChild(spacer);
     }
 
+    // Key display
     if (!isRoot) {
-      const ks = document.createElement('span');
-      ks.className = 'key';
-      ks.textContent = Array.isArray(parentObj) ? `[${key}]` : `"${key}":`;
-      row.appendChild(ks);
+      const keySpan = document.createElement('span');
+      keySpan.className = 'key';
+      keySpan.textContent = Array.isArray(parentObj) ? `[${key}]` : `"${key}":`;
+      row.appendChild(keySpan);
     }
 
+    // Value handling
     if (isObject) {
-      const tl = document.createElement('span');
-      tl.style.fontSize = '10px';
-      tl.style.opacity = '0.5';
-      tl.textContent = Array.isArray(value) ? `Array[${value.length}]` : '{Object}';
-      row.appendChild(tl);
+      const typeLabel = document.createElement('span');
+      typeLabel.className = 'type-label';
+      typeLabel.textContent = Array.isArray(value) ? `Array(${value.length})` : 'Object';
+      row.appendChild(typeLabel);
       
+      // Action buttons
       if (!this._isReadonly) {
-        const acts = document.createElement('div');
-        acts.className = 'actions';
-        const add = document.createElement('button');
-        add.className = 'btn';
-        add.textContent = '+';
-        add.onclick = () => {
-          if (Array.isArray(value)) value.push("");
-          else value[`new_${Object.keys(value).length}`] = "";
+        const actions = document.createElement('div');
+        actions.className = 'actions';
+        const addBtn = document.createElement('button');
+        addBtn.className = 'btn';
+        addBtn.textContent = '+';
+        addBtn.onclick = () => {
+          if (Array.isArray(value)) value.push(null);
+          else value[`newKey_${Object.keys(value).length}`] = null;
           this._refreshTree();
           this._dispatchChange();
         };
-        acts.appendChild(add);
-        row.appendChild(acts);
+        actions.appendChild(addBtn);
+        
+        const copyBtn = document.createElement('button');
+        copyBtn.className = 'btn';
+        copyBtn.textContent = '📋';
+        copyBtn.title = 'Copy JSON';
+        copyBtn.onclick = () => {
+          navigator.clipboard.writeText(JSON.stringify(value, null, 2))
+            .then(() => {
+              const originalText = copyBtn.textContent;
+              copyBtn.textContent = '✅';
+              setTimeout(() => copyBtn.textContent = originalText, 1000);
+            })
+            .catch(err => console.error('Failed to copy: ', err));
+        };
+        actions.appendChild(copyBtn);
+
+        if (!isRoot) {
+          const delBtn = document.createElement('button');
+          delBtn.className = 'btn btn-del';
+          delBtn.textContent = '×';
+          delBtn.onclick = () => {
+            Array.isArray(parentObj) ? parentObj.splice(key, 1) : delete parentObj[key];
+            this._refreshTree();
+            this._dispatchChange();
+          };
+          actions.appendChild(delBtn);
+        }
+        row.appendChild(actions);
+      }
+
+      wrapper.appendChild(row);
+
+      const childrenCont = document.createElement('div');
+      childrenCont.className = 'children';
+      Object.entries(value).forEach(([k, v]) => {
+        childrenCont.appendChild(this._createNode(k, v, value));
+      });
+      wrapper.appendChild(childrenCont);
+
+    } else {
+      // Primitive Value Input
+      const input = document.createElement('input');
+      const valType = typeof value;
+      input.className = `val-${valType}`;
+      input.value = value === null ? 'null' : value;
+      input.disabled = this._isReadonly;
+
+      input.onchange = (e) => {
+        let newVal = e.target.value;
+        // Basic auto-type conversion
+        if (newVal === 'true') newVal = true;
+        else if (newVal === 'false') newVal = false;
+        else if (newVal === 'null') newVal = null;
+        else if (!isNaN(newVal) && newVal.trim() !== '') newVal = Number(newVal);
+        
+        if (Array.isArray(parentObj)) parentObj[key] = newVal;
+        else parentObj[key] = newVal;
+        
+        this._dispatchChange();
+        this._refreshTree();
+      };
+      row.appendChild(input);
+
+      // Actions for primitives
+      if (!this._isReadonly) {
+        const actions = document.createElement('div');
+        actions.className = 'actions';
+
+        const copyBtn = document.createElement('button');
+        copyBtn.className = 'btn';
+        copyBtn.textContent = '📋';
+        copyBtn.title = 'Copy Value';
+        copyBtn.onclick = () => {
+          navigator.clipboard.writeText(String(value))
+            .then(() => {
+              const originalText = copyBtn.textContent;
+              copyBtn.textContent = '✅';
+              setTimeout(() => copyBtn.textContent = originalText, 1000);
+            })
+            .catch(err => console.error('Failed to copy: ', err));
+        };
+        actions.appendChild(copyBtn);
+
+        if (!isRoot) {
+          const delBtn = document.createElement('button');
+          delBtn.className = 'btn btn-del';
+          delBtn.textContent = '×';
+          delBtn.onclick = () => {
+            Array.isArray(parentObj) ? parentObj.splice(key, 1) : delete parentObj[key];
+            this._refreshTree();
+            this._dispatchChange();
+          };
+          actions.appendChild(delBtn);
+        }
+        row.appendChild(actions);
       }
       wrapper.appendChild(row);
-      const childs = document.createElement('div');
-      childs.className = 'children';
-      Object.entries(value).forEach(([k, v]) => childs.appendChild(this._createNode(k, v, value)));
-      wrapper.appendChild(childs);
-    } else {
-      const inp = document.createElement('input');
-      inp.value = value;
-      inp.disabled = this._isReadonly;
-      inp.onchange = (e) => {
-        let v = e.target.value;
-        if (v === 'true') v = true; else if (v === 'false') v = false; else if (!isNaN(v) && v !== '') v = Number(v);
-        parentObj[key] = v;
-        this._dispatchChange();
-      };
-      row.appendChild(inp);
-      wrapper.appendChild(row);
     }
+
     return wrapper;
   }
 
+  // API getters/setters
   get value() { return this._data; }
-  set value(v) { this._data = v; this._refreshTree(); }
+  set value(val) {
+    this._data = val;
+    this._refreshTree();
+  }
 });
 
-/**
- * <local-storage-editor>
- * Integrated with <vibe-json-editor> for graphical JSON manipulation.
- */
-class LocalStorageEditor extends HTMLElement {
-  static get observedAttributes() {
-    return [
-      'accent-color-hex',
-      'editor-height-pixels',
-      'ui-font-family',
-      'status-display-duration-ms'
-    ];
-  }
-
-  constructor() {
-    super();
-    this.attachShadow({ mode: 'open' });
-    this._activeJson = null;
-  }
-
-  connectedCallback() {
-    this.render();
-    this.refreshKeys();
-  }
-
-  attributeChangedCallback(name, oldValue, newValue) {
-    if (oldValue === newValue) return;
-    this.updateStyles();
-  }
-
-  refreshKeys() {
-    const select = this.shadowRoot.querySelector('#key-selector');
-    const currentSelection = select.value;
-    select.innerHTML = '<option value="" disabled selected>Select a key...</option>';
-    Object.keys(localStorage).sort().forEach(key => {
-      const opt = document.createElement('option');
-      opt.value = opt.textContent = key;
-      select.appendChild(opt);
-    });
-    if (currentSelection) select.value = currentSelection;
-  }
-
-  loadKey(key) {
-    const textarea = this.shadowRoot.querySelector('#editor-area');
-    const vibeEditor = this.shadowRoot.querySelector('#vibe-editor');
-    const rawValue = localStorage.getItem(key) || '';
-    
-    try {
-      const json = JSON.parse(rawValue);
-      this._activeJson = json;
-      vibeEditor.setAttribute('json-string-content', JSON.stringify(json));
-      vibeEditor.style.display = 'block';
-      textarea.style.display = 'none';
-    } catch (e) {
-      this._activeJson = null;
-      textarea.value = rawValue;
-      vibeEditor.style.display = 'none';
-      textarea.style.display = 'block';
-    }
-  }
-
-  saveData() {
-    const key = this.shadowRoot.querySelector('#key-selector').value;
-    const textarea = this.shadowRoot.querySelector('#editor-area');
-    const vibeEditor = this.shadowRoot.querySelector('#vibe-editor');
-
-    if (!key) return this.setStatus('Select a key', 'error');
-
-    let finalValue;
-    if (this._activeJson) {
-      finalValue = JSON.stringify(vibeEditor.value);
-    } else {
-      finalValue = textarea.value;
-    }
-
-    try {
-      localStorage.setItem(key, finalValue);
-      this.setStatus('Saved to Storage', 'success');
-      this.loadKey(key); // Refresh view
-    } catch (e) {
-      this.setStatus('Save Error', 'error');
-    }
-  }
-
-  setStatus(msg, type) {
-    const statusEl = this.shadowRoot.querySelector('#status-bar');
-    statusEl.textContent = msg;
-    statusEl.className = `status ${type}`;
-    const dur = parseInt(this.getAttribute('status-display-duration-ms')) || 3000;
-    setTimeout(() => statusEl.textContent = '', dur);
-  }
-
-  updateStyles() {
-    const host = this.shadowRoot.host;
-    host.style.setProperty('--accent', this.getAttribute('accent-color-hex') || '#6366f1');
-    host.style.setProperty('--editor-height', `${this.getAttribute('editor-height-pixels') || '400'}px`);
-    host.style.setProperty('--font-stack', this.getAttribute('ui-font-family') || 'system-ui, sans-serif');
-  }
-
-  render() {
-    this.shadowRoot.innerHTML = `
-      <style>
-        :host {
-          --accent: #6366f1; --bg: #ffffff; --border: #e2e8f0; --text: #1e293b;
-          --editor-height: 400px; --font-stack: system-ui, sans-serif;
-          display: block; max-width: 900px; margin: 1rem auto; font-family: var(--font-stack);
-          border: 1px solid var(--border); border-radius: 12px; background: var(--bg); overflow: hidden;
-          box-shadow: 0 10px 15px -3px rgb(0 0 0 / 0.1);
-        }
-        .container { padding: 1.5rem; display: flex; flex-direction: column; gap: 1rem; }
-        .toolbar { display: flex; gap: 0.6rem; align-items: center; }
-        select, textarea, button { font-family: inherit; font-size: 0.9rem; border-radius: 8px; border: 1px solid var(--border); padding: 0.6rem; }
-        select { flex: 1; }
-        textarea { width: 100%; height: var(--editor-height); background: #f8fafc; resize: vertical; box-sizing: border-box; font-family: monospace; }
-        #vibe-editor { 
-          height: var(--editor-height); overflow-y: auto; border: 1px solid var(--border); 
-          border-radius: 8px; padding: 10px; background: #fff; display: none;
-        }
-        button { cursor: pointer; font-weight: 600; background: #f1f5f9; transition: all 0.2s; border: none; }
-        button:hover { background: #e2e8f0; }
-        button.primary { background: var(--accent); color: white; }
-        .status { font-size: 0.8rem; font-weight: bold; }
-        .status.success { color: #10b981; }
-        .status.error { color: #ef4444; }
-        h2 { margin: 0; font-size: 1.1rem; color: var(--accent); }
-        .header { display: flex; justify-content: space-between; align-items: center; }
-      </style>
-      <div class="container">
-        <div class="header">
-          <h2>Storage Vibe Editor</h2>
-          <div id="status-bar" class="status"></div>
-        </div>
-        <div class="toolbar">
-          <select id="key-selector"></select>
-          <button id="refresh-btn">↻</button>
-          <button style="color:#ef4444" id="delete-btn">Delete</button>
-        </div>
-        
-        <vibe-json-editor id="vibe-editor"></vibe-json-editor>
-        <textarea id="editor-area" placeholder="Raw value..."></textarea>
-
-        <div class="toolbar">
-          <button class="primary" id="save-btn">Save Changes</button>
-          <button id="toggle-type-btn">Switch to Raw/JSON</button>
-        </div>
-      </div>
-    `;
-
-    this.updateStyles();
-    const sel = this.shadowRoot.querySelector('#key-selector');
-    sel.onchange = (e) => this.loadKey(e.target.value);
-    this.shadowRoot.querySelector('#refresh-btn').onclick = () => { this.refreshKeys(); this.setStatus('Refreshed', 'success'); };
-    this.shadowRoot.querySelector('#save-btn').onclick = () => this.saveData();
-    this.shadowRoot.querySelector('#delete-btn').onclick = () => {
-      if (confirm('Delete this key?')) {
-        localStorage.removeItem(sel.value);
-        this.refreshKeys();
-        this.loadKey('');
-      }
-    };
-    this.shadowRoot.querySelector('#toggle-type-btn').onclick = () => {
-      const v = this.shadowRoot.querySelector('#vibe-editor');
-      const t = this.shadowRoot.querySelector('#editor-area');
-      if (v.style.display === 'none') {
-        try {
-          this._activeJson = JSON.parse(t.value);
-          v.setAttribute('json-string-content', t.value);
-          v.style.display = 'block'; t.style.display = 'none';
-        } catch(e) { this.setStatus('Not valid JSON', 'error'); }
-      } else {
-        t.value = JSON.stringify(v.value, null, 2);
-        this._activeJson = null;
-        v.style.display = 'none'; t.style.display = 'block';
-      }
-    };
-
-    this.shadowRoot.querySelector('#vibe-editor').addEventListener('json-change', (e) => {
-      this._activeJson = e.detail.json;
-    });
-  }
-}
-
